% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilizr.R
\docType{data}
\name{atus}
\alias{atus}
\title{American Time Use Survey Data}
\format{
A data frame with 8,136 observations of 38 variables
}
\source{
\url{http://www.bls.gov/tus/}

\url{https://www.bls.gov/tus/lexicons/lexiconwex2022.pdf}
}
\usage{
data(atus)
}
\description{
A dataset containing responses from the 2022 American Time Use Survey. Each
participant is asked to document the amount of time spent participating in
various activities for a single 24 hour period. Variables provided are a
subset of the entire survey. The American Time Use Survey Activity Lexicon,
which is linked below, includes examples of different activities and how they
are sorted into categories.
}
\details{
\itemize{
\item age. age of respondent
\item sex. sex of respondent
\item education_level. education level of respondent
\item race. race of respondent
\item hispanic. whether the respondent is hispanic
\item metropolitan. whether the respondent lives in a metropolitan area
\item employed. status of employment
\item student. whether respondent is a student
\item spouse_presence. whether spouse is present in household of respondent
\item children_number. number of children of respondent
\item sleep. time spent sleeping, in minutes
\item groom. time spent grooming, in minutes
\item personal_care_misc. time spent on personal care other than sleep and groom, in minutes
\item hh_activities. household activities, in minutes
\item hh_member_care. caring for and helping household members, in minutes
\item nonhh_member_care. caring for and helping nonhousehold members, in minutes
\item work. time spent working, in minutes
\item class. time spent in class, in minutes
\item extracurricular. time spent on extracurricular activities, in minutes
\item homework. time spent on homework, in minutes
\item education_misc. time spent on education, other than class, extracurricular, and homework, in minutes
\item shopping. time spent shopping, in minutes
\item pro_services. time spent using professional and personal care services, in minutes
\item household_services. time spent using household_services, in minutes
\item government. time spent on government services and civic obligations, in minutes
\item eat_drink. time spent eating and drinking, in minutes
\item socialize. time spent socializing, in minutes
\item television. time spent watching television, in minutes
\item read. time spent reading, in minutes
\item leisure_misc. time spent on leisure besides socializing, television, and reading, in minutes
\item sports_participate. time spent participating in sports, in minutes
\item sports_spectate. time spent spectating sports, in minutes
\item sports_misc. time spent on sports, besides participating and spectating, in minutes
\item religious. time spent on religious activities, in minutes
\item volunteer. time spent volunteering, in minutes
\item telephone. time spent on telephone calls, in minutes
\item travel. time spent traveling, in minutes
\item unable_to_code. remaining time, which cannot be categorized into one of the previous categories, in minutes
}
}
\keyword{dataset}
