% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilizr.R
\docType{data}
\name{futbol}
\alias{futbol}
\title{Futbol Data Set}
\format{
A data frame with 57 observations of 6 variables
}
\source{
Data from: \url{http://www.profootballlogic.com/players/?position=all}, \url{https://www.ussoccer.com/teams/usmnt/roster}, \url{https://www.fifacm.com/}
}
\usage{
data(futbol)
}
\description{
Player data for National Football League 1st string quarterbacks from 2023 or from members of
the U.S. Men's National Soccer Team from 2023 Gold Cup roster
}
\details{
\itemize{ 
\item team. The home city for each player's team. 
\item player. The name of the player. 
\item ht_inches. The player's height in inches. 
\item wt_lbs. The player's weight in pounds. 
\item age. The age of the player in 2023. 
\item league. Indicates whether the player plays for the U.S. Men's National Soccer Team or the NFL.
}
}
\keyword{dataset}
