% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mobilizr.R
\docType{data}
\name{surveys}
\alias{surveys}
\title{Surveys Data}
\format{
A data frame with 36 observations of 4 variables
}
\usage{
data(surveys)
}
\description{
This synthetic dataset contains survey responses of customers of an imaginary clothing store.
}
\details{
\itemize{
\item customer_id. A number identifying each customer
\item age. Age of customer, in years
\item gender. M or F
\item favorite_product. Favorite clothing product of customer
}
}
\keyword{dataset}
